+++ src/admin/src/components/base/visibilityToggle/VisibilityToggle.stories.js
@@ -1,0 +1,6 @@
+/* eslint-disable */
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import VisibilityToggle from './VisibilityToggle';
+
+storiesOf('VisibilityToggle', module).add('default', () => <VisibilityToggle />);
