+++ src/admin/src/components/page/advisory/Advisory.css
@@ -16,9 +16,18 @@
 
 .ad-field {
   padding: 10px;
+  border-radius: 4px !important;
 }
 
 .ad-flex {
+  display: flex;
+}
+
+.ad-flex-wrap {
+  flex-wrap: wrap !important;
+}
+
+.ad-flex-date {
   display: flex;
 }
 
@@ -72,12 +81,17 @@
 }
 
 .ad-interval-select {
-  width: 50% !important;
-  margin-right: 3px !important;
+  width: 100% !important;
+  margin-left: 0px !important;
+}
+
+.ad-auto-margin {
+  margin: auto !important;
 }
 
 .MuiRadio-root,
-.MuiCheckbox-root {
+.MuiCheckbox-root,
+.MuiIconButton-root {
   color: inherit !important;
   padding: 0px !important;
 }
@@ -97,6 +111,11 @@
   color: #234075 !important;
 }
 
+[class*="-MenuList"] [class*="-option"],
+[class*="-singleValue"] {
+  color: #234075 !important;
+}
+
 .pointer {
   cursor: pointer;
 }
@@ -112,8 +131,5 @@
   .ad-btn-group button {
     width: 100% !important;
     margin-top: 5px !important;
-  }
-  .ad-interval-select {
-    width: 100% !important;
   }
 }
