+++ src/admin/src/components/page/advisory/Advisory.stories.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import CreateAdvisory from './CreateAdvisory';
+import CreateAdvisory from './Advisory';
 
 storiesOf('CreateAdvisory', module).add('default', () => <CreateAdvisory />);
