+++ src/admin/src/components/page/advisory/Advisory.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import CreateAdvisory from './CreateAdvisory';
+import CreateAdvisory from './Advisory';
 
 describe('<CreateAdvisory />', () => {
   test('it should mount', () => {
